+++ resources/views/layouts/app.blade.php
@@ -7,6 +7,15 @@
 
     <title>{{ config('app.name', 'Világműhely') }} - @yield('title', 'Workshop Crystals')</title>
 
+    <!-- Google Analytics -->
+    <script async src="https://www.googletagmanager.com/gtag/js?id=G-488487734"></script>
+    <script>
+        window.dataLayer = window.dataLayer || [];
+        function gtag(){dataLayer.push(arguments);}
+        gtag('js', new Date());
+        gtag('config', 'G-488487734');
+    </script>
+
     @vite(['resources/css/app.css', 'resources/js/app.js'])
     @stack('head')
 </head>